Renamed defaultcommand packages
