+++ defaultclicommands/dnscmd/Dns.go
@@ -1,4 +1,4 @@
-package dns
+package dnscmd
 
 import (
 	"fmt"
