+++ defaultclicommands/gitlabcmd/gitlabmetricscmd/ExposeMetrics.go
@@ -1,4 +1,4 @@
-package metrics
+package gitlabmetricscmd
 
 import (
 	"context"
