+++ defaultclicommands/monitoringcmd/prometheuscmd/ReadMetric.go
@@ -1,4 +1,4 @@
-package prometheus
+package prometheuscmd
 
 import (
 	"context"
